requirements.txt
fastapi
uvicorn
openai
langchain
faiss-cpu
python-dotenv
tiktoken